forgot password delete user results to deleting all the credit and debit transaction of the user git commit -m

+++ resources/user.py
@@ -1,21 +1,31 @@
 from flask_smorest import Blueprint, abort
 from flask.views import MethodView
 from passlib.hash import pbkdf2_sha256
-from schema import UserSchema, UserLoginschema, plainUserSchema, transactionByUser
+from schema import (
+    UserSchema,
+    UserLoginschema,
+    plainUserSchema,
+    transactionByUser,
+    TokenReset,
+)
 from model import User
-from extensions import db
+from extensions import db, mail
 from flask_jwt_extended import (
     create_access_token,
     create_refresh_token,
     get_jwt,
     get_jwt_identity,
-    jwt_required
+    jwt_required,
 )
 from blocklist import BLOCKLIST
 from sqlalchemy import or_
 from datetime import timedelta
+from flask_mail import Message
+import random
 
 blb = Blueprint("user", __name__, description="a user api")
+
+reset_token = random.randint(0000, 9999)
 
 
 @blb.route("/user/register")
@@ -36,7 +46,7 @@ class UserRegister(MethodView):
             email=user_data["email"],
             password=pbkdf2_sha256.hash(user_data["password"]),
             phone_number=user_data["phone_number"],
-            account_number=int(user_data["phone_number"])
+            account_number=int(user_data["phone_number"]),
         )
         db.session.add(user)
         db.session.commit()
@@ -49,8 +59,9 @@ class TokenRefresh(MethodView):
     @jwt_required(refresh=True)
     def post(self):
         current_user = get_jwt_identity()
-        new_token = create_access_token(identity=current_user,
-                                        expires_delta=timedelta(hours=2))
+        new_token = create_access_token(
+            identity=current_user, expires_delta=timedelta(hours=2)
+        )
 
         return {"access_token": new_token}
 
@@ -59,8 +70,7 @@ class TokenRefresh(MethodView):
 class UserLogin(MethodView):
     @blb.arguments(UserLoginschema)
     def post(self, user_data):
-        user = User.query.filter(
-                User.username == user_data["username"]).first()
+        user = User.query.filter(User.username == user_data["username"]).first()
 
         if user and pbkdf2_sha256.verify(user_data["password"], user.password):
             access_token = create_access_token(fresh=True, identity=user.id)
@@ -78,7 +88,7 @@ class UserLogin(MethodView):
     @jwt_required()
     def post(self):
         # get the current user's token
-        jti = get_jwt()['jti']
+        jti = get_jwt()["jti"]
         # send the token to the BLOCKLIST set in the blocklist.py file
         # this will revoke the token. A new access token will be created for you when you log in again
         BLOCKLIST.add(jti)
@@ -120,6 +130,8 @@ class GetSpecificUser(MethodView):
     @jwt_required()
     def delete(self, user_id):
         if User.query.get(get_jwt_identity()).is_admin:
+            if user_id == get_jwt_identity():
+                abort(401, message="Cannot delete yourself")
             user = User.query.get_or_404(user_id)
             db.session.delete(user)
             db.session.commit()
@@ -150,3 +162,47 @@ class GetUserTransaction(MethodView):
         current_user_id = get_jwt_identity()
         current_user = User.query.get(current_user_id)
         return current_user.transacts
+
+
+@blb.route("/forgot_password/<int:user_id>/<string:username>")
+class passwordForgot(MethodView):
+    def post(self, user_id, username):
+        try:
+            user_exist = User.query.get_or_404(user_id)
+            if user_exist:
+                if user_exist.username == username:
+                    msg = Message(
+                        "Reset Token",
+                        sender=user_exist.email,
+                        recipients=[
+                            user_exist.email
+                        ],
+                        )
+                    msg.body = f"Your reset token is: {reset_token}"
+                    mail.send(msg)
+                    return {"message": "Sent successfully"}
+                abort(401, message="Credentials does not match")
+            abort(404, message="User not found")
+        except:
+            abort(417, message="Check your data connection/credentials")
+
+
+@blb.route("/reset_password/<int:user_id>/<string:username>")
+class passwordReset(MethodView):
+    @blb.arguments(TokenReset)
+    def post(self, data, username, user_id):
+        res_token = data["res_token"]
+        password = data["password"]
+        confirm_password = data["confirm_password"]
+
+        if res_token != reset_token:
+            abort(401, message="Invalid token")
+        if confirm_password != password:
+            abort(401, message="Password does not match")
+
+        user = User.query.filter_by(username=username).first()
+        if user.id != user_id:
+            abort(401, message="Credentials does not match")
+        user.password = pbkdf2_sha256.hash(password)
+        db.session.commit()
+        return {"message": "Password has been reset, you can now login"}
